+++ Engine/Code/Engine/Renderer/RenderSceneGraph.cpp
@@ -25,7 +25,7 @@ void RenderSceneGraph::AddParticleEmitter( ParticleEmitter* p ) {
 void RenderSceneGraph::RemoveRenderable( Renderable* r) {
 	std::vector<Renderable*>::iterator searchResult = std::find(m_renderables.begin(), m_renderables.end(), r);
 	if (searchResult != m_renderables.end()) {
-		unsigned int index = searchResult - m_renderables.begin();
+		unsigned int index = (unsigned int) (searchResult - m_renderables.begin());
 		m_renderables[index] = m_renderables[ m_renderables.size() - 1 ];
 		m_renderables.pop_back();
 	}
@@ -35,7 +35,7 @@ void RenderSceneGraph::RemoveRenderable( Renderable* r) {
 void RenderSceneGraph::RemoveCamera( Camera* c ) {
 	std::vector<Camera*>::iterator searchResult = std::find(m_cameras.begin(), m_cameras.end(), c);
 	if (searchResult != m_cameras.end()) {
-		unsigned int index = searchResult - m_cameras.begin();
+		unsigned int index = (unsigned int) (searchResult - m_cameras.begin());
 		m_cameras[index] = m_cameras[ m_cameras.size() - 1 ];
 		m_cameras.pop_back();
 	}
@@ -45,7 +45,7 @@ void RenderSceneGraph::RemoveCamera( Camera* c ) {
 void RenderSceneGraph::RemoveLight( Light* l ) {
 	std::vector<Light*>::iterator searchResult = std::find(m_lights.begin(), m_lights.end(), l);
 	if (searchResult != m_lights.end()) {
-		unsigned int index = searchResult - m_lights.begin();
+		unsigned int index = (unsigned int) (searchResult - m_lights.begin());
 		m_lights[index] = m_lights[ m_lights.size() - 1 ];
 		m_lights.pop_back();
 	}
@@ -56,7 +56,7 @@ void RenderSceneGraph::RemoveLight( Light* l ) {
 void RenderSceneGraph::RemoveParticleEmitter( ParticleEmitter* l ) {
 	std::vector<ParticleEmitter*>::iterator searchResult = std::find(m_particleEmitters.begin(), m_particleEmitters.end(), l);
 	if (searchResult != m_particleEmitters.end()) {
-		unsigned int index = searchResult - m_particleEmitters.begin();
+		unsigned int index = (unsigned int) (searchResult - m_particleEmitters.begin());
 		RemoveRenderable(m_particleEmitters[index]->renderable);
 		m_particleEmitters[index] = m_particleEmitters[ m_particleEmitters.size() - 1 ];
 		m_particleEmitters.pop_back();
@@ -71,11 +71,11 @@ void RenderSceneGraph::SortCameras() {
 
 
 unsigned int RenderSceneGraph::GetRenderableCount() const {
-	return m_renderables.size();
+	return (unsigned int) m_renderables.size();
 }
 unsigned int RenderSceneGraph::GetLightCount() const {
-	return m_lights.size();
+	return (unsigned int) m_lights.size();
 }
 unsigned int RenderSceneGraph::GetCameraCount() const {
-	return m_cameras.size();
+	return (unsigned int) m_cameras.size();
 }
